+++ header/Net/MQTTHandler.h
@@ -16,7 +16,7 @@ namespace Net
 
 		static UInt32 __stdcall KAThread(void *userObj);
 	public:
-		MQTTHandler(Net::SocketFactory *sockf, const Net::SocketUtil::AddressInfo *addr, UInt16 port, const UTF8Char *username, const UTF8Char *password, Net::MQTTClient::PublishMessageHdlr hdlr, void *userObj, UInt32 kaSeconds);
+		MQTTHandler(Net::SocketFactory *sockf, const Net::SocketUtil::AddressInfo *addr, UInt16 port, const UTF8Char *username, const UTF8Char *password, Net::MQTTClient::PublishMessageHdlr hdlr, void *userObj, UInt16 kaSeconds);
 		~MQTTHandler();
 
 		Bool IsError();
