+++ src/Media/Resizer/LanczosResizer16_C8_CPP.cpp
@@ -26,7 +26,7 @@ extern "C" void LanczosResizer16_C8_horizontal_filter_pa(UInt8 *inPt, UInt8 *out
 			cvals = PSADDW4(cvals, PLoadInt16x4(&rgbTable[ReadUInt16(&currIn[0]) * 8 + 1310720]));
 			cvals = PSADDW4(cvals, PLoadInt16x4(&rgbTable[ReadUInt16(&currIn[2]) * 8 + 786432]));
 			cvals = PSADDW4(cvals, PLoadInt16x4(&rgbTable[ReadUInt16(&currIn[4]) * 8 + 262144]));
-			cvals = PMULM2HW4(cvals, PInt16x4SetA(ReadUInt16(&currIn[6]) >> 1));
+			cvals = PMULM2HW4(cvals, PInt16x4SetA((Int16)(ReadUInt16(&currIn[6]) >> 1)));
 			cvals = PSADDW4(cvals, PLoadInt16x4(&rgbTable[ReadUInt16(&currIn[6]) * 8 + 1835008]));
 			PStoreInt16x4(tmpPtr, cvals);
 			currIn += 8;
@@ -185,7 +185,7 @@ extern "C" void LanczosResizer16_C8_expand_pa(UInt8 *inPt, UInt8 *outPt, OSInt w
 			cvals = PLoadInt16x4(&rgbTable[1310720 + ReadUInt16(&inPt[0]) * 8]);
 			cvals = PSADDW4(cvals, PLoadInt16x4(&rgbTable[786432 + ReadUInt16(&inPt[2]) * 8]));
 			cvals = PSADDW4(cvals, PLoadInt16x4(&rgbTable[262144 + ReadUInt16(&inPt[4]) * 8]));
-			cvals = PMULM2HW4(cvals, PInt16x4SetA(ReadUInt16(&inPt[6]) >> 1));
+			cvals = PMULM2HW4(cvals, PInt16x4SetA((Int16)(ReadUInt16(&inPt[6]) >> 1)));
 			cvals = PSADDW4(cvals, PLoadInt16x4(&rgbTable[1835008 + ReadUInt16(&inPt[6]) * 8]));
 			PStoreInt16x4(outPt, cvals);
 			inPt += 8;
@@ -265,7 +265,7 @@ extern "C" void LanczosResizer16_C8_imgcopy_pa(UInt8 *inPt, UInt8 *outPt, OSInt 
 			cvals = PLoadInt16x4(&rgbTable[1310720 + ReadUInt16(&inPt[0]) * 8]);
 			cvals = PSADDW4(cvals, PLoadInt16x4(&rgbTable[786432 + ReadUInt16(&inPt[2]) * 8]));
 			cvals = PSADDW4(cvals, PLoadInt16x4(&rgbTable[262144 + ReadUInt16(&inPt[4]) * 8]));
-			cvals = PMULM2HW4(cvals, PInt16x4SetA(ReadUInt16(&inPt[6]) >> 1));
+			cvals = PMULM2HW4(cvals, PInt16x4SetA((Int16)(ReadUInt16(&inPt[6]) >> 1)));
 			cvals = PSADDW4(cvals, PLoadInt16x4(&rgbTable[1835008 + ReadUInt16(&inPt[6]) * 8]));
 			ucvals = PCONVI16x4_U(cvals);
 			outPt[0] = rgbTable[0 + PEXTUW4(ucvals, 0)];
